feat(pipeline): Move entity processing into pipeline
Move entity processing steps from web/query into pipeline.

Processing extensions is still done in web/query since moving this
requires rewriting quite a number of tests.
