+++ tests/datasets/test_sessions_processing.py
@@ -1,6 +1,5 @@
 from snuba.clickhouse.query import Query
 from snuba.datasets.factory import get_dataset
-from snuba.pipeline.single_query_plan_pipeline import SingleQueryPlanPipeline
 from snuba.query import SelectedExpression
 from snuba.query.expressions import Column, CurriedFunctionCall, FunctionCall, Literal
 from snuba.query.parser import parse_query
@@ -46,14 +45,6 @@ def test_sessions_processing() -> None:
         ]
         return QueryResult({}, {})
 
-    pipeline = (
-        sessions.get_default_entity()
-        .get_query_pipeline_builder()
-        .build_pipeline(request, query_runner)
-    )
-    assert isinstance(pipeline, SingleQueryPlanPipeline)
-    query_plan = pipeline.query_plan
-
-    query_plan.execution_strategy.execute(
-        query_plan.query, request.settings, query_runner
-    )
+    sessions.get_default_entity().get_query_pipeline_builder().build_pipeline(
+        request, query_runner
+    ).execute()
